A File System developed in Unix (with Qemu)
The file system was developed for the subject "Expansion of Operating Systems". Components: it has developed the distribution of blocks (superblock, inode storage, root directory and readme) for data storage. It has implemented the functions corresponding to reading and writing data, updating and editing the information stored in the inodes, etc.